Fix lambda parameters
